Extracted FormBaseValidator
