+++ corehq/apps/app_manager/views/forms.py
@@ -64,7 +64,7 @@ from corehq.apps.domain.decorators import (
     track_domain_request,
     LoginAndDomainMixin,
 )
-from corehq.apps.app_manager.const import USERCASE_PREFIX, USERCASE_TYPE
+from corehq.apps.app_manager.const import USERCASE_PREFIX, USERCASE_TYPE, WORKFLOW_FORM
 from corehq.apps.app_manager.dbaccessors import get_app, get_apps_in_domain
 from corehq.apps.app_manager.helpers.validators import load_case_reserved_words
 from corehq.apps.app_manager.models import (
@@ -81,7 +81,6 @@ from corehq.apps.app_manager.models import (
     IncompatibleFormTypeException,
     ModuleNotFoundException,
     UpdateCaseAction,
-    WORKFLOW_FORM,
     CustomAssertion,
     CustomInstance,
     CaseReferences,
